Add files via upload
Changes! I've gotten rid of a lot of little syntax bugs, fixed the backdrop for the webpage. In addition, Prettier was throwing several strange errors my way that I had to look into- I spent a lot of time Googling and found issues with my Javascript, information about nesting tags (the errors that I was getting dealt with a div tag that was accidentally placed inside of a p-tag) and general clean-up.
The carousel is coming along, but is still proving more work than I had anticipated. All the same, I march ever closer.